Adjust z-index of elements in a more compreehensive way
Default elements across the whole portal that need the use of z-index now use values in increments of 10. Any customization of z-index values inside any of these major default elements should use values between them.
Eg. If custom z-index values are now needed in elements inside the edit-bar, these should have values between 11 and 19, which is before the default value of 20 for the #column-one, and after the value of 10 for the edit-bar. Values in actual content elements should be below 10, like the ones used in carrousel. And so goes.